[tests] TestVideoInfoScanner: fix sign-compare warning

+++ xbmc/video/test/TestVideoInfoScanner.cpp
@@ -69,6 +69,6 @@ TEST(TestVideoInfoScanner, EnumerateEpisodeItemByTitle)
   CFileItem item("/foo.special.mp4", false);
   EPISODELIST result;
   ASSERT_TRUE(scanner.EnumerateEpisodeItem(&item, result));
-  ASSERT_EQ(result.size(), 1);
+  ASSERT_EQ(result.size(), static_cast<size_t>(1));
   ASSERT_EQ(result[0].strTitle, "foo");
 }
